Add files via upload
Question:
A farmer is on the east side of a river with one wolf, two cats and three ducks. The wolf will
start a fight with any cats it shares a side with, unless the cats outnumber the wolf. Similarly,
any cats on the same side as a duck will start a fight, unless there are more ducks than cats.
But if the farmer is on the same side, he will prevent any fight before it starts.
The farmer’s boat can carry up to two animals at a time, in any combination. He is permitted
to cross the river without any animals in his boat (unless doing so would cause a fight to
break out!)
How can the farmer get all his animals to the other side without a fight? 
A state is represented as a list containing four elements. The first element represents the number
of wolves on the east bank; the second represents the number of cats on the east bank; the third
represents the number of ducks on the east bank; the fourth represents the location of the farmer,
which can be east or west. Thus, under this representation the initial state would be represented
as (1 2 3 east).